prod-ocr/pyproject.toml
[project]
name = "prod-ocr"
version = "0.1.0"
description = "AI-powered document classification and data extraction MVP using Google Gemini API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "pypdf",
    "python-dotenv",
    "azure-functions",
    "azure-storage-blob",
    "azure-storage-queue",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]

[tool.hatch.build.targets.sdist]
include = [
    "modules",
    "function_app.py",
    "send_task.py",
    "get_results.py",
    "README.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "responses>=0.23.0",
]
